ui-stuff/audit-dashboard-ui-main/requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Database and data handling (sqlite3 is built into Python)
pandas>=2.0.0
numpy>=1.24.0

# File handling and processing
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0

# Email and notifications
email-validator>=2.0.0

# Security and authentication
bcrypt>=4.0.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# AI and ML (for Claude AI integration)
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
